requirements.txt
absl-py==1.0.0
anyio==3.4.0
apturl==0.5.2
asgiref==3.4.1
astunparse==1.6.3
attrs==20.3.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.8.2
cached-property==1.5.2
cachetools==4.2.4
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
command-not-found==0.3
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.5.0
distro-info==1.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.17
fastapi==0.70.0
fasteners==0.14.1
flatbuffers==2.0
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h11==0.12.0
h5py==3.6.0
httplib2==0.18.1
httptools==0.2.0
idna==2.10
importlib-metadata==4.8.2
jeepney==0.6.0
jsonschema==3.2.0
keras==2.7.0
Keras-Preprocessing==1.1.2
keyring==22.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
libclang==12.0.0
lockfile==0.12.2
louis==3.16.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.9
numpy==1.21.4
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
paramiko==2.7.2
pexpect==4.8.0
Pillow~=8.4.0
pipreqs==0.4.11
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==2.0.1
pydantic==1.8.2
PyGObject==3.38.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt===2.2.0-ubuntu0.21.04.1
python-dateutil==2.8.1
python-debian==0.1.39
python-dotenv==0.19.2
pytz==2021.1
pyxdg==0.27
PyYAML==5.3.1
reportlab==3.5.66
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.8
SecretStorage==3.3.1
simplejson==3.17.2
six==1.15.0
sniffio==1.2.0
starlette==0.16.0
systemd-python==234
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
texttable==1.6.3
typing-extensions==4.0.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.2
usb-creator==0.3.7
uvicorn==0.15.0
uvloop==0.16.0
wadllib==1.3.5
watchgod==0.7
websocket-client==0.57.0
websockets==10.1
Werkzeug==2.0.2
wrapt==1.13.3
xdg==5
xkit==0.0.0
yarg==0.1.9
zipp==1.0.0

gym~=0.21.0
matplotlib~=3.5.1
